### 2024-08-23 #7138 @editorialbot
[PRE REVIEW]: PySLSQP: A transparent Python package for the SLSQP optimization algorithm modernized with utilities for visualization and post-processing
**Submitting author:** @anugrahjo (<a href="http://orcid.org/0009-0003-7704-2532">Anugrah Jo Joshy</a>)
**Repository:** https://github.com/anugrahjo/PySLSQP
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
