### 2024-08-23 #7138 @danielskatz
👋 @anugrahjo - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. (In this case, you have an invalid DOI. If there is no DOI for an author, remove the DOI line.) In addition, you could work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect, as well as fixing the DOIs with the extra prefix. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, your paper is 2700 words long, and JOSS papers are intended to be 250-1000 words, so you will have to substantially cut your paper. Again, see the example paper and the JOSS documentation for what a JOSS paper should be.

Finally, looking at the code itself, it appears the Fortran code here is not an original contribution, but the Python code is. Is this correct?
