+++ paper/paper.bib
@@ -15,6 +15,7 @@
   pages={262--281},
   year={1994},
   publisher={ACM New York, NY, USA}
+  doi= {10.1145/192115.192124}
 }
 
 @article{gill2005snopt,
@@ -26,7 +27,7 @@
   pages     = {99--131},
   year      = {2005},
   publisher = {SIAM},
-  doi       = {https://doi.org/10.1137/S0036144504446096}
+  doi       = {10.1137/S0036144504446096}
 }
 
 @article{virtanen2020scipy,
@@ -38,7 +39,7 @@
   pages={261--272},
   year={2020},
   publisher={Nature Publishing Group},
-  doi       = {https://doi.org/10.1038/s41592-019-0686-2}
+  doi= {10.1038/s41592-019-0686-2}
 }
 
 @article{wu2020pyoptsparse,
@@ -49,7 +50,7 @@
   number={54},
   pages={2564},
   year={2020},
-  doi={https://doi.org/10.21105/joss.02564}
+  doi={10.21105/joss.02564}
 }
 
 @article{perez2012pyopt,
@@ -60,7 +61,7 @@
   pages={101--118},
   year={2012},
   publisher={Springer},
-  doi={https://doi.org/10.1007/s00158-011-0666-3}
+  doi={10.1007/s00158-011-0666-3}
 }
 
 @misc{NLopt,
