### 2024-08-23 1b26c55
Fix a typo in the bib file.
