+++ paper/paper.bib
@@ -3,7 +3,7 @@
   title={A software package for sequential quadratic programming},
   author={Kraft, Dieter},
   journal={Forschungsbericht- Deutsche Forschungs- und Versuchsanstalt fur Luft- und Raumfahrt},
-  year={1988}
+  year={1988},
 }
 
 @article{kraft1994algorithm,
@@ -14,8 +14,8 @@
   number={3},
   pages={262--281},
   year={1994},
-  publisher={ACM New York, NY, USA}
-  doi= {10.1145/192115.192124}
+  publisher={ACM New York, NY, USA},
+  doi= {10.1145/192115.192124},
 }
 
 @article{gill2005snopt,
@@ -27,7 +27,7 @@
   pages     = {99--131},
   year      = {2005},
   publisher = {SIAM},
-  doi       = {10.1137/S0036144504446096}
+  doi       = {10.1137/S0036144504446096},
 }
 
 @article{virtanen2020scipy,
@@ -39,7 +39,7 @@
   pages={261--272},
   year={2020},
   publisher={Nature Publishing Group},
-  doi= {10.1038/s41592-019-0686-2}
+  doi= {10.1038/s41592-019-0686-2},
 }
 
 @article{wu2020pyoptsparse,
@@ -50,7 +50,7 @@
   number={54},
   pages={2564},
   year={2020},
-  doi={10.21105/joss.02564}
+  doi={10.21105/joss.02564},
 }
 
 @article{perez2012pyopt,
@@ -61,19 +61,19 @@
   pages={101--118},
   year={2012},
   publisher={Springer},
-  doi={10.1007/s00158-011-0666-3}
+  doi={10.1007/s00158-011-0666-3},
 }
 
 @misc{NLopt,
     author = {Steven G. Johnson},
     title  = {The {NLopt} nonlinear-optimization package},
     url    = {http://github.com/stevengj/nlopt},
-    year   = {2024}
+    year   = {2024},
 }
 
 @misc{modopt,
     author = {Anugrah Jo Joshy},
     title  = {A MODular development environment and library for OPTimization algorithms},
     url    = {https://github.com/LSDOlab/modopt},
-    year   = {2024}
+    year   = {2024},
 }
