### 2024-08-23 #7138 @anugrahjo
👋 @danielskatz Thank you for the feedback. I have made the changes to the references and the paper compiles now.

Sorry about the length of the paper. One of the JOSS papers in our list of references was of similar length so we thought the 250-1000 word length was a suggestion. We can reduce the length, if necessary.

Regarding the code contribution, yes, the Python code is the main contribution and the build workflows for autogenerating the precompiled binaries for the wrapped Fortran code. The Fortran code for SLSQP itself is open-source and in our case, it is sourced from the SciPy repository but is also available from many other sources.
