### 2024-08-23 #7138 @danielskatz
Paper length is one of many things in JOSS that has gotten stricter over time...

Please let me know when you have a shorter paper.  I'll mark this as paused for now.
