### 2024-08-23 #7138 @danielskatz
this command includes checking the length of the paper
