### 2024-08-23 #7138 @danielskatz
@anugrahjo - before you do any extra work, I'm also going to ask the JOSS editors to take a look at this for scope, since the commit history here and in https://github.com/anugrahjo/PySLSQP_alpha just goes back about 3 months.  There should be a decision on this in a week or two.
