### 2024-08-23 #7138 @anugrahjo
@danielskatz - I want to let you know that even the PySLSQP_alpha repository was created only after the software reached a reasonably stable working state. There is an earlier private repository that predates PySLSQP_alpha, as well as weeks of local development prior to the creation of the initial repository.
