### 2024-08-23 34b3431
Upload the first correctly compiled paper on JOSS and its preprint files (.tex and .pdf).
