### 2024-08-28 9c6ec2c
Update citation format
