### 2024-09-18 #7246 @prashjha
Dear @hariharanragothaman and @saaikrishnan, please read the first couple of comments in this thread and create your review checklist. You can read the reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Also, you can browse the [closed "REVIEW" issues](https://github.com/openjournals/joss-reviews/issues?q=label%3Areview+is%3Aclosed) on the "joss-reviews" repository to get some ideas on how to complete the reviews. Good luck!
