### 2024-10-07 #7246 @hariharanragothaman
@saaikrishnan Yes I agree.  ✅  There were 2 failing tests.

I think this is the reason for it.

1. The last deployment to pypi happened 5 months ago, whereas recent code updates (2 months ago)  in the repository have not been deployed to pypi. This implies the public packages has not been updated.

_Reference_:

- https://github.com/anugrahjo/PySLSQP/deployments
- Skipped deployment in github actions: https://github.com/anugrahjo/PySLSQP/actions/runs/10589035773/job/29342676773
