### 2024-10-08 #7246 @anugrahjo
@saaikrishnan
As @hariharanragothaman commented, if you are installing a release from PyPI but using the latest commit from the GitHub repo, the tests might not work. **The tests need to correspond to the PyPI release**. To test correctly, you can either
    1. install using `pip install pyslsqp` and run tests after cloning the repository and checking out the correct version with `git checkout v0.1.0`, or
    2. install the most recent commit with `pip install git+https://github.com/anugrahjo/PySLSQP.git@main` and run tests after cloning the latest commit.

I just verified locally that no tests are failing when either of these is followed. You can also see from the GitHub actions page that the tests on the cloud are successful for both the latest commit and the PyPI-released version. Let me know if you have any concerns.

Thanks.
