### 2024-10-08 #7246 @anugrahjo
@saaikrishnan @hariharanragothaman
Also, the failing tests you see are coming from the docstrings as there were minor changes made to how the console output is printed. These are only cosmetic changes that do not justify a new version release. We only deploy new versions of the package to PyPI which is why the latest commits on GitHub do not necessarily correspond to the latest PyPI release.
