### 2024-10-08 e52d1fe
Look recursively into windows build files
