+++ .github/workflows/build_install_test.yml
@@ -124,10 +124,15 @@ jobs:
         python -m pip install --upgrade pip
         python -m pip install pytest
         python -m pip install -e .  # Install the package in editable mode, "pip install ." did not seem to copy the compiled files to the site-packages directory
-        # cat meson_builddir/meson-logs/meson-log.txt
-        ls .
+        echo "----------------------------------------"
+        cat meson_builddir/meson-logs/meson-log.txt
+        echo "----------------------------------------"
+        ls -R .
+        echo "----------------------------------------"
         ls meson_builddir
+        echo "----------------------------------------"
         ls meson_builddir/pyslsqp
+        echo "----------------------------------------"
         ls pyslsqp
 
     # - name: Download Dependencies Tool
