### 2024-10-09 e0d7dd0
Add manual f2py compilation for Windows github actions
