### 2024-10-09 928eb2a
Set install=Trur for _slsqp module to see if it helps with Windows
