### 2024-10-09 f09dd58
Run f2py manually in setup.py for Windows to generate extension files (.pyd and .dll)
