### 2024-10-09 35d07b6
Update setup.py for Windows case
