+++ setup.py
@@ -79,6 +79,11 @@ if __name__ == "__main__":
 
         # Workaround for meson build failing to produce .dll on Windows
         if platform.system() == "Windows":
+            if not "FC" in os.environ:
+                os.environ["FC"] = "gfortran"
+            if not "CC" in os.environ:
+                os.environ["CC"] = "gcc"
+                
             original_dir = os.getcwd()
             os.chdir('pyslsqp/slsqp')
             subprocess.run(['python', '-m', 'numpy.f2py', '-c', 'slsqp.pyf', 'slsqp_optmz.f'], check=True)
