### 2024-10-10 591a232
Update build-system requires so that numpy is not affected by distutils deprecation.
