### 2024-10-10 b58e463
Set backend as meson for Windows during manual compilation.
