### 2024-10-10 9d95476
Update windows workflow in actions.
