+++ .github/workflows/build_install_test.yml
@@ -131,16 +131,16 @@ jobs:
         # ls -R .
         # echo "----------------------------------------"
         python -m pip install -e .  # Install the package in editable mode, "pip install ." did not seem to copy the compiled files to the site-packages directory
-        echo "----------------------------------------"
-        cat meson_builddir/meson-logs/meson-log.txt
-        echo "----------------------------------------"
-        ls -R .
-        echo "----------------------------------------"
-        ls meson_builddir
-        echo "----------------------------------------"
-        ls meson_builddir/pyslsqp
-        echo "----------------------------------------"
-        ls pyslsqp
+        # echo "----------------------------------------"
+        # cat meson_builddir/meson-logs/meson-log.txt
+        # echo "----------------------------------------"
+        # ls -R .
+        # echo "----------------------------------------"
+        # ls meson_builddir
+        # echo "----------------------------------------"
+        # ls meson_builddir/pyslsqp
+        # echo "----------------------------------------"
+        # ls pyslsqp
 
     # - name: Download Dependencies Tool
     #   run: |
