### 2024-10-10 b62ca74
List all files in the windows workflow in actions.
