### 2024-10-10 f579c52
add meson builddir for Windows manual build
