+++ setup.py
@@ -86,7 +86,7 @@ if __name__ == "__main__":
 
             original_dir = os.getcwd()
             os.chdir('pyslsqp/slsqp')
-            subprocess.run(['python', '-m', 'numpy.f2py', '-c', 'slsqp.pyf', 'slsqp_optmz.f', '--backend', 'meson'], check=True)
+            subprocess.run(['python', '-m', 'numpy.f2py', '-c', 'slsqp.pyf', 'slsqp_optmz.f', '--backend', 'meson', '--build-dir', 'meson_builddir'], check=True)
             os.chdir(original_dir)
 
             build_path  = os.path.join(os.getcwd(), 'pyslsqp', 'slsqp')
