### 2024-10-10 3490975
Replace time.time() with time.perf_counter() for more resolution in Windows machines and avoiding division by zero errors.
