### 2024-10-11 464a09f
Revert back to previous configurations for meson.build, pyproject.toml files. Only changes from v0.1.0 is in setup.py and build_install_test.yml for Windows.
