+++ pyslsqp/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '0.1.0'
+__version__ = '0.1.1-beta'
 
 from pyslsqp.main import optimize, get_default_options
