+++ pyproject.toml
@@ -1,5 +1,6 @@
 [build-system]
-requires = ['meson-python', 'setuptools', 'ninja', 'numpy<=1.26', "meson>=1.1", "wheel"]
+requires = ['meson-python', 'setuptools', 'ninja', 'numpy']
+# requires = ['meson-python', 'setuptools', 'ninja', 'numpy<=1.26', "meson>=1.1", "wheel"]
 build-backend = 'setuptools.build_meta'
 
 [project]
