### 2024-10-11 c23f910
Remove py38 for Windows manual testing. Add actions workflow for building Windows 2019 and 2022 wheels.
