### 2024-10-11 d1f17f8
Remove py38 for Windows manual testing and all wheel building.
