+++ .github/workflows/build_wheels_test_deploy.yml
@@ -64,13 +64,19 @@ jobs:
           # To silence this warning: set `CIBW_TEST_SKIP: "cp38-macosx_*:arm64"`
           CIBW_TEST_SKIP: "cp38-macosx_arm64"
 
-      - name: Build linux and windows wheels
-        if: false == contains( matrix.os, 'macos')
+      - name: Build linux wheels
+        if: contains( matrix.os, 'ubuntu')
         uses: pypa/cibuildwheel@v2.18.1
 
         # Uncomment the following 2 lines for faster testing of the workflow
         # env:
         #   CIBW_BUILD: "cp38-* cp312-*"
+
+      - name: Build Windows wheels # only for python>=3.9
+        if: contains( matrix.os, 'windows')
+        uses: pypa/cibuildwheel@v2.18.1
+        env:
+          CIBW_SKIP: "cp38-*"
 
       - uses: actions/upload-artifact@v4
         with:
@@ -107,7 +113,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-22.04, macos-13, macos-14, windows-2019, windows-2022]
-        python-version: ["3.8", "3.12"]
+        python-version: ["3.9", "3.12"]
     
     if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags/v')
     steps:      
