### 2024-10-11 b918cff
Correctly skip CIBW for cp38 in Windows using pyproject.toml.
