+++ .github/workflows/build_wheels_test_deploy.yml
@@ -75,8 +75,9 @@ jobs:
       - name: Build Windows wheels # only for python>=3.9
         if: contains( matrix.os, 'windows')
         uses: pypa/cibuildwheel@v2.18.1
-        env:
-          CIBW_SKIP: "cp38-*"
+        # env:
+        #   CIBW_SKIP: "cp38-* *-win32" # win32 needs to be specified since this overrides skip specified in pyproject.toml 
+        #                               # Now handled in pyproject.toml windows subsection
 
       - uses: actions/upload-artifact@v4
         with:
