### 2024-10-11 bbd1c49
Correctly skip CIBW for cp38 in Windows using pyproject.toml.
