+++ pyproject.toml
@@ -87,4 +87,4 @@ test-requires = "pytest"
 test-command = 'pytest {project}/tests -m "not visualize"'
 
 [tool.cibuildwheel.windows]
-skip = ['cp38-*']
+skip = ["cp38-*", "*-win32"]
