### 2024-10-11 d4a0380
Update pyproject for cibw.windows to prevent the error "no module named numpy" when trying to build numpy.f2py extension on the cloud by running the subprocess in setup.py specific to windows. Revert back to msys2 from rtools.
