+++ pyproject.toml
@@ -88,3 +88,4 @@ test-command = 'pytest {project}/tests -m "not visualize"'
 
 [tool.cibuildwheel.windows]
 skip = ["cp38-*", "*-win32"]
+before-build = "pip install wheel numpy ninja setuptools meson meson-python wheel"
