### 2024-10-11 203de1d
Update pyproject for cibw.windows to build with no-isolation
