### 2024-10-11 c26b002
Run "pip install numpy" in CIBW_BEFORE_ALL for Windows wheel builds on cloud.
