### 2024-10-13 9d534bc
Revert back changes in the last 3 commits for fixing "Module numpy not found" while building Windows wheels in setup.py subprocess. Use sys.executable to run f2py compilation in the subprocess call using the same Python interpreter and environment.
