+++ .github/workflows/build_wheels_test_deploy.yml
@@ -81,8 +81,7 @@ jobs:
       - name: Build Windows wheels # only for python>=3.9
         if: contains( matrix.os, 'windows')
         uses: pypa/cibuildwheel@v2.18.1
-        env:
-          CIBW_BEFORE_ALL: pip install numpy
+        # env:
         #   CIBW_SKIP: "cp38-* *-win32" # win32 needs to be specified since this overrides skip specified in pyproject.toml 
         #                               # Now handled in pyproject.toml windows subsection
 
