### 2024-10-13 9bc3fca
Remove inclusion of paper/ directory in sdist
