+++ MANIFEST.in
@@ -6,5 +6,8 @@ recursive-include docs *
 # Exclude the _build directory from the docs in the source distribution
 recursive-exclude docs/_build *
 
+# Exclude the paper directory in the source distribution
+recursive-exclude paper *
+
 # Exclude any .so files from the source distribution
 global-exclude *.so
