### 2024-10-13 014910e
Delete the meson_builddir/ in slsqp/ folder after running f2py so that it won"t be copied in sdist and cause any issues while building Windows wheels.
