### 2024-10-13 516b9c0
Fix a critical issue in package_data in ste=up.py
