+++ setup.py
@@ -110,7 +110,7 @@ if __name__ == "__main__":
 
     setup(
         include_package_data=True,
-        package_data={'pyslsqp': ["*.so", ".lib", ".pyd", ".pdb", ".dylib", ".dll"]}, # this is needed to include the shared object file in the build directory in site-pkgs
+        package_data={'pyslsqp': ["*.so", "*.lib", "*.pyd", "*.pdb", "*.dylib", "*.dll"]}, # this is needed to include the shared object file in the build directory in site-pkgs
         # cmdclass={'bdist_wheel': bdist_wheel},  # The cmdclass argument in the setup function is used to override default commands provided by setuptools.
         #                                         # This overrides the `bdist_wheel` command. BUT THIS IS SPECIFIC TO WHEELS DISTRIBUTION FORMAT ONLY.
         #                                         # The `bdist_wheel` command is used by setuptools to build a wheel distribution of our package. 
