### 2024-10-13 19df924
Skip building PyPy wheels for Windows architectures in CIBW
