### 2024-10-13 a86d377
Update release notes and changelog for new release.
