+++ release_notes.md
@@ -1,5 +1,13 @@
+# PySLSQP 0.1.1 (October 13, 2024)
+
+- Add support for Windows OS (cloud tested and wheels made available on PyPI) 
+- More precise timing with perf_counter()
+- Fix swapped printing of average function and derivative evaluation times in final results
+- Minor updates to documentation
+- Add PySLSQP software paper to the repository
+
 ***********************************
-# PySLSQP 0.1.0 (JUNE 2, 2024)
+# PySLSQP 0.1.0 (June 2, 2024)
 Initial Release of PySLSQP.
 
 Summary of major changes from the previous release - None
