+++ docs/src/changelog.md
@@ -6,6 +6,15 @@ in [`RELEASE_NOTES.md`](https://github.com/anugrahjo/PySLSQP/blob/main/release_n
 in the project root directory.
 
 ***********************************
+## Version 0.1.1 (October 13, 2024)
+
+- Add support for Windows OS (cloud tested and wheels made available on PyPI) 
+- More precise timing with perf_counter()
+- Fix swapped printing of average function and derivative evaluation times in final results
+- Minor updates to documentation
+- Add PySLSQP software paper to the repository
+
+***********************************
 ## Version 0.1.0 (June 2, 2024)
 
 - Initial release of PySLSQP
