### 2024-10-13 24469f7
Correct workflow for installing and testing sdist on Windows
