+++ paper/paper.md
@@ -45,7 +45,7 @@ that wraps the original Fortran code sourced from the SciPy repository and provi
 a host of new features to improve the research utility of the original algorithm.
 `PySLSQP` uses a simple yet modern workflow for compiling and using Fortran code from Python.
 This allows even beginner developers to easily modify the algorithm in Fortran
-for their purposes and use in Python the wrapper auto-generated by the workflow.
+for their specific needs and use in Python the wrapper auto-generated by the workflow.
 
 Some of the additional features offered by `PySLSQP` include auto-generation of
 unavailable derivatives using finite differences, independent scaling of the problem 
@@ -124,11 +124,11 @@ The total number of function evaluations is indicated within parentheses in the 
 We see that `PySLSQP` is the only algorithm that solves the problem
 within the 200 function evaluation limit.
 Among the algorithms that failed to converge are SNOPT, TrustConstr, and IPOPT.
-SNOPT is a commercial SQP algorithm, while TrustConstr and IPOPT are Interior Point (IP) algorithms.
+SNOPT is a commercial SQP algorithm, while TrustConstr and IPOPT are open-source Interior Point (IP) algorithms.
 Although IPOPT appears to have converged in the plot, the solution returned by IPOPT 
 does not satisfy the feasibility criteria.
 This underscores the relevance of SLSQP even today among state-of-the-art optimization algorithms.
-This problem is taken from the suite of examples in the modOpt [@modopt] optimization library.
+This problem is taken from the suite of examples in the modOpt [@joshy2024modopt] optimization library.
 
 
 There are several optimization libraries in Python that include the SLSQP algorithm,
@@ -162,8 +162,8 @@ scaling, warm/hot restarting, and post-processing.
 
 By addressing the current limitations and providing new capabilities, `PySLSQP` enhances the transparency and usability
 of the SLSQP algorithm, making it a more powerful and user-friendly tool for solving nonlinear programming problems.
-`PySLSQP` is now interfaced with the modOpt [@modopt] library of optimizers, through which it has successfully
-solved problems in aircraft design, as well as aircraft and spacecraft optimal control.
+`PySLSQP` is now integrated with the modOpt [@joshy2024modopt] library of optimizers, through which it has successfully
+solved problems in aircraft design, spacecraft optimal control, and swimming robot design.
 
 # Software features
 
@@ -247,18 +247,18 @@ In contrast, during a warm start, although the algorithm starts from the previou
 the Hessian is initialized as the identity matrix, which may necessitate more iterations to achieve convergence.
 
 ## Ease of extension
-`PySLSQP` is implemented in Python and relies on NumPy's *f2py* and the *Meson* build system 
+`PySLSQP` is implemented in Python and leverages NumPy's *f2py* and the *Meson* build system 
 for compiling and interfacing the underlying Fortran code with Python.
 The Python setup script automates the build process, making it straightforward for
-developers to build, install, and use `PySLSQP` after making modifications to the Fortran code.
-The package also includes GitHub workflows for automatically generating precompiled binaries 
+developers to build, install, test, and use `PySLSQP` after modifying the Fortran code.
+The package includes GitHub workflows to automatically generate precompiled binaries 
 in the cloud for different system architectures using PyPA's *cibuildwheel* tool.
 These automated workflows ensure that `PySLSQP` remains accessible to a broad range of users 
-by providing consistent and reliable installation across different platforms. 
+by providing consistent and reliable installation across various platforms using Linux, macOS, and Windows operating systems.
 Additionally, this approach allows developers to focus on enhancing the core algorithm 
 and features without the overhead of managing complex build environments, 
 thus fostering an open-source community that can contribute effectively 
[…]
