+++ paper/paper.bib
@@ -77,3 +77,11 @@
     url    = {https://github.com/LSDOlab/modopt},
     year   = {2024},
 }
+
+@article{joshy2024modopt,
+  title={modOpt: A modular development environment and library for optimization algorithms},
+  author={Joshy, Anugrah Jo and Hwang, John T},
+  journal={arXiv preprint arXiv:2410.12942},
+  year={2024},
+  doi={10.48550/arXiv.2410.12942}
+}
