+++ .github/workflows/build_wheels_test_deploy.yml
@@ -161,7 +161,7 @@ jobs:
       - name: Install the sdist
         run: |
           python -m pip install --upgrade pip
-          # python -m pip install ./dist/*.tar.gz  # This wildcard exapansion does not work on Windows
+          # python -m pip install ./dist/*.tar.gz  # This wildcard expansion does not work on Windows
           # Use Python’s glob module to explicitly handle the wildcard in a platform-independent way:
           python -m pip install $(python -c "import glob; print(glob.glob('./dist/*.tar.gz')[0])")
           python -m pip install pytest
