### 2024-11-25 e15d6b5
Allow numpy 2.x in pyproject.toml
