+++ pyproject.toml
@@ -18,7 +18,7 @@ maintainers = [
 keywords=['slsqp', 'optimization', 'optimization algorithm', 'nlp', 'optimizer', 'nonlinear programming', 'numerical optimization']
 dynamic = ["version"]
 dependencies = [
-  "numpy>=1.24,<2", 
+  "numpy>=1.24,<3", 
   "h5py>=3.11.0,<4", 
   "matplotlib>=3.4,<4",
 ]
