### 2024-11-25 bee370c
Use np.asarray(, dtype=np.float64) instead of np.asfarray() to support numpy 2.0
