### 2024-11-25 3045ff6
Remove expired np.float_ to support numpy 2.0
