### 2024-11-25 b6912c3
Update docstrings to include printing differences in numpy 2.x.x.
