### 2024-11-25 e224973
Update docstrings for numpy 2.x.x.
