### 2024-11-25 c427ede
Fix docstrings causing doctest failures.
