### 2024-11-25 6f67fc6
Release version 0.1.2 with NumPy 2.x support.
