+++ docs/src/changelog.md
@@ -6,7 +6,7 @@ in [`RELEASE_NOTES.md`](https://github.com/anugrahjo/PySLSQP/blob/main/release_n
 in the project root directory.
 
 ***********************************
-# Version 0.1.2 (November 24, 2024)
+## Version 0.1.2 (November 24, 2024)
 
 - Add support for NumPy 2.x.x
 - Update PySLSQP software paper
