### 2024-11-25 #7246 @anugrahjo
Hi @prashjha,

I updated the code today and released a new version 0.1.2. The latest version's files are archived and available on Zenodo at https://zenodo.org/records/14215102.
