### 2024-11-29 #7246 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
