### 2024-11-30 #7246 @anugrahjo
> @anugrahjo, I used v0.1.2 instead of 0.1.2 as v0.1.2 matches the tag in your repository. If you want 0.1.2, let me know, and I will change it.

@prashjha v0.1.2 sounds good.
