+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: '`PySLSQP`: A transparent Python package for the SLSQP optimization algorithm 
+title: 'PySLSQP: A transparent Python package for the SLSQP optimization algorithm 
 modernized with utilities for visualization and post-processing'
 # modernized with utilities for scaling, recording, restarting, visualization, and post-processing'
 tags:
@@ -14,12 +14,10 @@ authors:
     # corresponding: true
     affiliation: 1
   - name: John T. Hwang
-    affiliation: 2
+    affiliation: 1
 affiliations:
- - name: PhD Candidate, Department of Mechanical and Aerospace Engineering, University of California San Diego
+ - name: Department of Mechanical and Aerospace Engineering, University of California San Diego, USA
    index: 1
- - name: Associate Professor, Department of Mechanical and Aerospace Engineering, University of California San Diego
-   index: 2
 date: 6 August 2024
 bibliography: paper.bib
 ---
@@ -39,13 +37,13 @@ solving a sequence of Quadratic Programming (QP) subproblems.
 The Sequential Least SQuares Programming algorithm, or SLSQP, has been one of the 
 most widely used SQP algorithms since the 1980s.
 
-We present `PySLSQP`, a seamless interface for using the SLSQP algorithm from Python,
+We present `PySLSQP`, a seamless interface for using the SLSQP algorithm from Python
 <!-- The `PySLSQP` package provides a seamless interface for using the SLSQP algorithm from Python. -->
 that wraps the original Fortran code sourced from the SciPy repository and provides 
 a host of new features to improve the research utility of the original algorithm.
 `PySLSQP` uses a simple yet modern workflow for compiling and using Fortran code from Python.
 This allows even beginner developers to easily modify the algorithm in Fortran
-for their specific needs and use in Python the wrapper auto-generated by the workflow.
+for their specific needs and use, in Python, the wrapper auto-generated by the workflow.
 
 Some of the additional features offered by `PySLSQP` include auto-generation of
 unavailable derivatives using finite differences, independent scaling of the problem 
