+++ paper/paper.bib
@@ -7,7 +7,7 @@
 }
 
 @article{kraft1994algorithm,
-  title={Algorithm 733: TOMP--Fortran modules for optimal control calculations},
+  title={Algorithm 733: TOMP--{F}ortran modules for optimal control calculations},
   author={Kraft, Dieter},
   journal={ACM Transactions on Mathematical Software (TOMS)},
   volume={20},
@@ -31,7 +31,7 @@
 }
 
 @article{virtanen2020scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {P}ython},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
@@ -43,7 +43,7 @@
 }
 
 @article{wu2020pyoptsparse,
-  title={pyOptSparse: A Python framework for large-scale constrained nonlinear optimization of sparse systems},
+  title={pyOptSparse: A {P}ython framework for large-scale constrained nonlinear optimization of sparse systems},
   author={Wu, Neil and Kenway, Gaetan and Mader, Charles A and Jasa, John and Martins, Joaquim RRA},
   journal={Journal of Open Source Software},
   volume={5},
@@ -54,7 +54,7 @@
 }
 
 @article{perez2012pyopt,
-  title={pyOpt: a Python-based object-oriented framework for nonlinear constrained optimization},
+  title={pyOpt: a {P}ython-based object-oriented framework for nonlinear constrained optimization},
   author={Perez, Ruben E and Jansen, Peter W and Martins, Joaquim RRA},
   journal={Structural and Multidisciplinary Optimization},
   volume={45},
