### 2024-11-30 #7246 @danielskatz
@anugrahjo - I'm suggesting a few changes in your paper in https://github.com/anugrahjo/PySLSQP/pull/4

The change to the title is because we don't use special formatting in titles.

The changes to the affiliation are because JOSS affiliations are just places, not positions.

I also note that this paper is far too long for JOSS. We discussed this when the submission happened, but it doesn't appear that you've addressed it. I suppose this is partially my fault for not enforcing this when the review started. I won't make you do anything about it now.  (@prashjha - as a note, you also can enforce shortening papers of this length in the future, even if the track editor forgets.)
